requirements.txt
Django==4.2.7
mysqlclient==2.2.0
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-simple-captcha==0.6.0
openpyxl==3.1.2
reportlab==4.0.7
python-decouple==3.8

